# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/control/force-spin/{tableId}": {
      "post": {
        "operationId": "ControlController_forceSpin",
        "parameters": [
          {
            "name": "tableId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
    "/api/v1/admin/control/logs": {
      "get": {
        "operationId": "ControlController_getLogs",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
    "/api/v1/admin/control/create-promo": {
      "post": {
        "operationId": "ControlController_createPromo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
    "/api/v1/admin/control/update": {
      "post": {
        "operationId": "ControlController_updateControl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateControlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
    "/api/v1/admin/control/categories": {
      "get": {
        "operationId": "ControlController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      },
      "delete": {
        "operationId": "ControlController_deleteCategory",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
    "/api/v1/admin/control/promotions/create": {
      "post": {
        "operationId": "ControlController_createPromotion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
    "/api/v1/admin/control/promotions": {
      "get": {
        "operationId": "ControlController_findAllPromotion",
        "parameters": [
          {
            "name": "group",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
    "/api/v1/admin/control/promotions/{id}": {
      "get": {
        "operationId": "ControlController_findOnePromotion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      },
      "patch": {
        "operationId": "ControlController_updatePromotion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromotionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      },
      "delete": {
        "operationId": "ControlController_removePromotion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Control"
        ]
      }
    },
...
    "/api/v1/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/admin-login": {
      "post": {
        "operationId": "AuthController_adminLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API documentation for backend services",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreatePromoDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateControlDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePromotionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePromotionDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}